server/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Claude Agent SDK
claude-agent-sdk>=0.1.0

# HTTP client for health checks
httpx>=0.26.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
